Separate calculate module from utils
