+++ src/utils/convert.ts
@@ -1,5 +1,3 @@
-import { Point, VisualizerMachineContext } from "../machines/visualizerMachine";
-
 export const convertDegreeToRadian = (angleInDegrees: number) => {
   return (Math.PI * angleInDegrees) / 180;
 };
@@ -10,21 +8,4 @@ export const convertToPercent = (ratio: number) => {
 
 export const convertToRatio = (percent: number) => {
   return percent / 100;
-};
-
-export const convertToViewportPoint = ({
-  devicePixelRatio,
-  canvasPoint,
-  zoom,
-  origin,
-}: {
-  devicePixelRatio: number;
-  canvasPoint: Point;
-  zoom: VisualizerMachineContext["zoom"];
-  origin: VisualizerMachineContext["origin"];
-}) => {
-  return {
-    x: (canvasPoint.x * zoom + origin.x) / devicePixelRatio,
-    y: (canvasPoint.y * zoom + origin.y) / devicePixelRatio,
-  };
 };
